BC-530 added more tests
